Update packages name
